backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.2.2
openpyxl==3.1.5
pydantic==2.5.0
python-dotenv==1.0.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# File extraction dependencies
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0

# AI/ML for classification
anthropic==0.34.2

# Utilities
aiofiles==23.2.1
